# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VulnLab Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      operationId: unfollowUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── PRODUCTS ─────────────────────────────────────────────────
...
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required: [ok]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        full_name:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
        balance:
          type: ["number", "null"]
        ssn:
          type: ["string", "null"]
        secret_note:
          type: ["string", "null"]
        password_plain:
          type: ["string", "null"]
        is_active:
          type: boolean
        is_locked:
          type: boolean
        failed_logins:
          type: integer
        last_login:
          type: ["string", "null"]
        created_at:
          type: string
        followers_count:
          type: ["integer", "null"]
        following_count:
          type: ["integer", "null"]
      required: [id, username, role, is_active, is_locked, failed_logins, created_at]
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
        bio:
          type: string
        phone:
          type: string
        address:
          type: string
        avatar_url:
          type: string
        role:
          type: string
...